refactor(renderer/api/tools): move interfaces to co-located types/ folder
